Refactor: Update Tailwind & config; migrate Next UI to HeroUI
