+++ package.json
@@ -11,8 +11,7 @@
     "format": "prettier --write ."
   },
   "dependencies": {
-    "@heroui/react": "^2.6.14",
-    "@nextui-org/react": "^2.6.11",
+    "@heroui/react": "2.7.4",
     "framer-motion": "^12.4.2",
     "next": "15.1.7",
     "ofetch": "^1.4.1",
@@ -34,10 +33,10 @@
     "eslint": "^9",
     "eslint-config-next": "15.1.7",
     "eslint-plugin-react": "^7.37.4",
-    "postcss": "8.4.49",
+    "postcss": "^8.5.3",
     "prettier": "^3.5.1",
     "prettier-plugin-tailwindcss": "^0.6.11",
-    "tailwindcss": "3.4.17",
+    "tailwindcss": "3",
     "typescript": "^5"
   }
 }
