+++ src/app/components/FormContainer.tsx
@@ -1,4 +1,4 @@
-import { Card, CardBody, CardHeader } from "@nextui-org/react";
+import { Card, CardBody, CardHeader } from "@heroui/react";
 
 interface FormContainerProps {
     title: string;
